refactor: move LoRA scaling tests into test_trainer.py
Move TestLoRaScaling class from test_trainer_utils.py into
test_trainer.py as suggested in review, and revert
test_trainer_utils.py to its original state.
